# lib/api-zod/src/generated/types/addChildInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tether API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddChildInput {
  displayName: string;
  pin: string;
  grade?: string;
  age?: number;
  avatarColor?: string;
}